sermon_audio/pyproject.toml
[tool.poetry]
name = "sermon-audio"
version = "0.1.0"
description = ""
authors = ["Zack Warren <warr7020@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sermonaudio = "^6.15.2"
google-api-python-client = "^2.122.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
